requirements.txt
# FastAPI and dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database and storage
qdrant-client>=1.1.1
neo4j>=4.4.0
aioneo4j>=0.0.5

# Memory system
mem0>=0.1.23
numpy>=1.21.0
openai>=1.0.0

# WebSocket support
websockets>=10.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Utilities
pytz>=2021.1
python-dotenv>=0.19.0
aiohttp>=3.8.0
tenacity>=8.0.1
httpx>=0.23.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
pytest-mock>=3.6.1
httpx>=0.23.0
asgi-lifespan>=1.0.1

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
pylint>=2.9.6
